+++ rand_distr/src/binomial.rs
@@ -13,6 +13,8 @@ use crate::{Distribution, Uniform};
 use rand::Rng;
 use core::fmt;
 use core::cmp::Ordering;
+#[allow(unused_imports)]
+use num_traits::Float;
 
 /// The binomial distribution `Binomial(n, p)`.
 ///
@@ -28,7 +30,7 @@ use core::cmp::Ordering;
 /// let v = bin.sample(&mut rand::thread_rng());
 /// println!("{} is from a binomial distribution", v);
 /// ```
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 pub struct Binomial {
     /// Number of trials.
@@ -344,5 +346,10 @@ mod test {
     #[should_panic]
     fn test_binomial_invalid_lambda_neg() {
         Binomial::new(20, -10.0).unwrap();
+    }
+
+    #[test]
+    fn binomial_distributions_can_be_compared() {
+        assert_eq!(Binomial::new(1, 1.0), Binomial::new(1, 1.0));
     }
 }