+++ rand_distr/src/geometric.rs
@@ -3,6 +3,8 @@
 use crate::Distribution;
 use rand::Rng;
 use core::fmt;
+#[allow(unused_imports)]
+use num_traits::Float;
 
 /// The geometric distribution `Geometric(p)` bounded to `[0, u64::MAX]`.
 /// 
@@ -25,7 +27,7 @@ use core::fmt;
 /// let v = geo.sample(&mut rand::thread_rng());
 /// println!("{} is from a Geometric(0.25) distribution", v);
 /// ```
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 pub struct Geometric
 {
@@ -232,5 +234,10 @@ mod test {
         let variance =
             results.iter().map(|x| (x - mean) * (x - mean)).sum::<f64>() / results.len() as f64;
         assert!((variance - expected_variance).abs() < expected_variance / 10.0);
+    }
+
+    #[test]
+    fn geometric_distributions_can_be_compared() {
+        assert_eq!(Geometric::new(1.0), Geometric::new(1.0));
     }
 }