+++ rand_pcg/src/lib.rs
@@ -38,7 +38,9 @@
 #![no_std]
 
 mod pcg128;
+mod pcg128cm;
 mod pcg64;
 
 pub use self::pcg128::{Lcg128Xsl64, Mcg128Xsl64, Pcg64, Pcg64Mcg};
+pub use self::pcg128cm::{Lcg128CmDxsm64, Pcg64Dxsm};
 pub use self::pcg64::{Lcg64Xsh32, Pcg32};